Title: 십진수와 이진수, Time: -1ms, Memory: -1MB, Status: Compile Error - CodeTree

+++ 240317/십진수와 이진수/decimal-and-binary-number.java
@@ -4,7 +4,7 @@ public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         char[] arr = sc.next().toCharArray();
-        int num = 0;
+        long num = 0;
         for(int i = 0 ; i < arr.length ; i++){
             num += (arr[i] - '0') * Math.pow(2, arr.length - 1 - i);
         }
